player Health, rough SpawnAtFlag

+++ Assets/Scripts/Events/TakeDamage.cs
@@ -3,6 +3,6 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
-public class TakeDamage : UnityEvent<float>
+public class TakeDamage : UnityEvent<int>
 {
 }
